Optimize sanitization methods in STPCardValidator
Strings that are already numeric can now be sanitized with zero
allocations, and strings that need sanitization will now be sanitized
faster. In addition, `+stringIsNumeric:` will never allocate.

In a simple benchmark, calling `+sanitizedNumericStringForString:` on
the string `"4242 4242 4242 4242"` is now twice as fast as before.

This also changes the sanitization methods to only allow ASCII digits.
The previous use of `+decimalDigitCharacterSet` allowed any character
in the Decimal Digit unicode category, such as Indic and Arabic decimal
digits.