Make "state" address field only required in US
If you request full billing info, but are not in the United States, do not require the state/county/province field (as many countries do not have something applicable).

Will revisit in the future how much per-country custom logic we want here, but this will at least make it functional in other countries.