Fix bin range testing.
Previously partial numbers did not match to any ranges (eg 6 did not match to 622). This meant that the validator sometimes incorrectly said things were invalid when they were in fact incomplete and would also not correctly return `brandForNumber:` for these partial matches. Changed this logic so partial numbers match bin ranges correctly and updated tests.

Also added a bunch of tests for hypthetical bin ranges which start with 0, which I expect is the most likely sort of future bug in this type of numeric comparison.