Change zip validation logic.
Valid zip if numeric string is 5 or 9 chars.
Previously potentiallyValid required only numerals, but isValid sanitized the string so there was an inconsistency. Now potentiallyValid is sanitizedNumeric <= 9 which is consistent (but maybe not exactly right because then something like 123A45 is a valid zip? Should revisit logic at some point)