Fix double-view dismissal
Hitting the done button in SFSafariVC makes it dismiss itself, so we don't want to manually dismiss it in that case because that will cause two dismissals to be fired and that can cascade to dismiss other view controllers. Instead, only dismiss in the other completion cases.

Also remove an uncessary completion block in the dismiss logic.

Fixes https://github.com/stripe/stripe-ios/issues/717