Fix a bunch of layout issues
Bump minimum padding
More exactly size cvc and expiry when not first responder so layout looks spaced evenly.
Size cvc more exactly to the max possible length when brand is known.
Fix which views are allowed to be hidden for cvc and postal.
Don't directly trigger a layoutSubviews from within the animation block (as this seems to have negative consequences). Moved layout logic to new function which is now called from both layoutSubviews and the animation block.