Add source tuple tests
Adds a bunch of tests for parsing source tuples with customers in various configurations.
Also:
Adds constants for the file names of many json test files.
Adds a customer fixture builder function that allows you to pass in arbitrary json source keys to build a customer with those sources in that order.