Adding STPAddress `containsContent...ForFields:` methods to check for any useful data
These are like an `isEmpty` check for the address, but they check the fields needed
for a particular billing or shipping address setting.

This addresses #831, which was introduced when
`STPShippingAddressViewController.billingAddress` started receiving an empty STPAddress
from STPCard instead of a nil one.

I had trouble deciding on the semantics of `containsContentForFields: .None` (should it
return true or false?). I chose to always return NO. If the question is "Does this address
have useful data for me, and I'm interested in none of the data from this address?", then
false/NO makes sense. It also parallels with the behavior when messaging nil.

Unit tests to come, and I'm not in love with the method names.