Fix (dumb) test failures.
I made a dumb mistake in STPAddCardViewController, declared the wrong type for the
`requiredFields` variable.

The snapshot tests for Add Card and Shipping Address were apparently relying on the
using a non-nil & empty Address to display the "Use <> address" button. Put a piece of
data into the address to keep triggering the button.