Fix an STPURLCallbackListener retain cycle.
Previously there was a bug where STPURLCallbackListener retained its listeners, which meant that STPRedirectContext had a retain cycle and would not deallocate while it was actively listening for callbacks. Fixing this is in some ways a breaking API change.
This commit fixes the bug, and also updates the docs to explicitly mention that you need to retain the instances you make.
It also replaces referencing to STPAPIClient source polling (which is deprecated) and instead links to the sources best practices doc.