Add `addressFieldTableViewCountryCode` to the `STPAddress` created by `STPAddressViewModel`
When there's no country table cell/text field, the country information is stored
in `addressFieldTableViewCountryCode`. This fixes #840, because the address used to
check `isValid` is created through this method.

I think this is a desirable change overall. When `setAddress:` is called, it extracts the
`STPAddress.country` value, and stores it into `addressFieldTableViewCountryCode`. It
makes a lot of sense to me to reverse that when reading the `STPAddress` back out of the
object.