Fix bugs around `STPPaymentCardTextField becomeFirstResponder`
Previous implementation of `nextFirstResponderField` would return `nil` if all fields were
valid, which blocked programatic calls to `becomeFirstResponder`.

It *also* used to cycle through the fields every time it was called, which doesn't seem
right. It feels more natural for the `nextFirstResponderField` to be either the first
invalid field *or* the last field. This puts the insertion point at the right place
to start deleting.

It also feels better if the user manually changes to the expiration field before entering
the card number: once the expiration is valid the firstResponder jumps back to numbers.
Then, once they complete numbers, it jumps to the CVC.

I also added a test for this behavior. Some of the things it verifies:

* becomeFirstResponder should not change the current first responder if the current field isn't valid yet
* becomeFirstResponder should take the user to the first invalid field when called
* becomeFirstResponder should make the last sub-field first responder when all the fields are valid