Fix regression: `nextFirstResponderField` *should* prefer to go to the next field, even if others are invalid
It's good to have the explicit progression through fields. Keep the previous behavior that
when they're on the last field and finish filling it out, take them back to the first
invalid field.

Also, update `canBecomeFirstResponder` and `becomeFirstResponder` to *never* change which
field is the first responder, if one of the subfields already is. If there isn't one,
use either the first invalid field or the last field.

Updates tests to reflect the new expected behavior.