Correctly enable Done button after copying billing address into shipping (& vice-versa) (#871)
Trigger `addressFieldTableViewCellDidUpdateText:` after `setContents:` in STPAddressFieldTableViewCell

I missed a bug around the "Use Shipping" and "Use Billing" buttons: the Done button wasn't
properly updated if the copied information was all that was needed to be complete.

This was introduced by #867. Prior to #867, *all* changes when through the delegate
method `formTextFieldTextDidChange:`

These buttons programmatically set the contents, and we still need the delegate method
to be called in that case.