Make STPEncodable work *through* arrays.
It previously only worked if the array contains primitive items (or dictionaries?).

Also, we need to provide array indices (which work with the Stripe API), to avoid ambiguity.
It won't work with just `[]`, because this particular test case could either deserialize
as a single object with two properties filled out, or two objects with one property
each.