Add Faux Pas annotations to `NSDateComponents<STPFormEncodable>` conformance.
Faux Pas (correctly) points out adding un-prefixed methods to a Foundation class via
a category is dangerous. Probably a better thing to do is either prefix all of these
methods in the protocol, or add custom support for `NSDateComponents` in
`+[STPFormEncoder formEncodableValueForObject:]`.

For now, I think this is safe enough.