Turn off userInteractionEnabled on `maskView`
This view controls what part of the text field is visible, and so we need to be able to
tap the text field behind it when it exists (happens on 4" devices)

In iOS 11, this view appears in the view hierarchy (seen in Xcode Visual Debugger), and
it's capturing touches. Tapping/clicking directly on the card numbers does not cause the
numbers text field to become first responder, instead it dismisses first responder!

This change fixes that, and now you can go back to the numbers field from the month field.