Make taps on the brandImage on payment card text field move focus to the numbers field.
Especially on 4" devices (in compact mode), there's a very small touch target for
activating the numbers field. The brand image is just as big, but taps on it will resign
first responder.

I considered doing something even more complicated: ideally any touch in the whole
STPPaymentCardTextField's bounds would give it first responder, and it would be the text
field that's closest to where the user tapped.

For this quick change, I don't really want to adjust the subfield bounds, or
change/understand the code that calculates the frames to increase their effective touch
targets.

Just making the brandImage tappable is a quick and easy win.