Rename "card.io" section to "Card IO"
We have a link to the "card.io" section with the id="cardio" but jazzy and GitHub generate a different id for the "card.io" section. Jazzy calls it "#cardio" whereas GitHub calls it "#card-io".

I looked into opening a PR to change the jazzy behavior to match the GitHub behavior but it looked pretty in depth (jazzy depends on a separate markdown gem) so just rename the section to avoid the problem.