Add empty localization files to Custom Integration example
This example doesn't use any localized strings but adding empty translations does enable the SDK to translate certain strings to the language.

For example, if the SDK is translated in French and the device is set to French, the French SDK localization won't be used unless the app itself claims that it can support French.