Workaround for SafariVC reporting failure while redirecting to iDEAL (#937)
Ignores reported errors from SafariVC unless we know the error was for a `stripe.com`
domain.

SafariVC is basically a black box: give it a URL and ask it to load. It'll report via
the delegate method when the initial load finishes, and whether it was successful or not.

Prior to this commit, iDEAL (and others?) would fail and close SafariVC. However, it hadn't
actually failed, and just leaving the SafariVC open would allow the user to finish
the redirect workflow.

It's strange to me that SafariVC behaves this way. It's also a little strange how the
redirects are being handled by girogate.de. This works around the interaction between
the two, while still attempting to report legitimate errors, so the host application can
tell the user and allow them to retry.

* Add error handling to `STPRedirectContext` completion block errors in example code.
* Rename property and move comment for code review feedback
* Add changelog entry