Make iDEAL `name` parameter optional, and accept empty string as nil. (#940)
When iDEAL support was originally implemented in the iOS SDK, this was a required
parameter. It isn't any more.

I considered removing the `name:` parameter completely, but this maintains backwards
compatibility.

Also handling the empty string the same as `nil` for these optional fields, so that there
isn't a surprise waiting. IMO, an optional field that contains the empty string shouldn't
cause the request to fail.