Prevent `STPPaymentMethodTableViewCell.titleLabel` from overlapping checkmark (#952)
Previous layout code allowed the `titleLabel` to be as wide as it wanted, even if
that ran offscreen. Fixing it to explicitly constrain the `titleLabel` to be between
the `leftIcon` and `checkmarkIcon`, with horizontal spacing of `padding`.

Instead of asking UIKit to calculate the height that the `titleLabel` should be, just use
the full height of the cell. We get the same vertical centering, and similarly prevents
the label from exceeding the bounds of the cell.

The snapshot tests do show slight differences in how the `titleLabel` renders. I believe
it's *now* pixel aligning the text. When I compare before/after, especially on something
like the `E` in `Ending In` (it's true of all the letters, but really apparent on `E`),
the old snapshot has anti-aliased top/bottoms, where the new snapshot has crisp lines.
This is subtle enough that I hadn't noticed it previously, but I think it's a nice,
unexpected bonus.

Fixes #948