Fixes crash in `STPColorUtils.perceivedBrightnessForColor` (#954)
This is used for deciding whether to use light/dark scroll bars & status bar indicators,
and choosing which color should be used for the switch in `STPSwitchTableViewCell`.

This code assumed the color was always in the RGB color space, which is incorrect. It would
reliably crash with Address Sanitization turned on, and give inconsistent answers with
it turned off.

Now using the `UIColor` `getRed:green:blue:alpha:` method to convert the color to RGB
before calculating the luma value. Also added some links to documentation for why this
calculation is interesting.

Added tests for `STPColorUtils colorIsBright:`
This also tests `perceivedBrightnessForColor:`, but does so by checking something that's
more meaningful to me, rather than testing for specific luma numbers