Add `dispatch_async` to `[STPRedirectContext handleWillForegroundNotification]` to allow `[STPRedirectContext handleURLCallback]` to be prioritized (#962)
We considered some other approaches and landed on this one:

Option 1:

Delete `STPURLCallback` completely

The downside to this approach is the host app will still get `[AppDelegate application:openURL:options:]` calls. If `STPURLCallback` was gone, the host app won't know whether or not that `[AppDelegate application:openURL:options:]` call should be ignored by their own app's routing logic.

Option 2:
Don't unschedule `STPURLCallback` and let the `STPRedirectContext` stay in the `STPRedirectContextStateCompleted` state

The downside to this approach is the `STPRedirectContext` ends up holding a non-negligible amount of memory in order to just store the `STPRedirectContextStateCompleted` state.

Option 3:

Similarly to the last one, we can avoid unscheduling the `STPURLCallback` for a second or so.

This is very similar to the approach we landed on but would require more significant code changes. Also the reasoning behind how much delay to have is unclear.