Update card expiration years to 2024 to fix failing tests (#983)
* Add assert with friendly-ish error message for a case that already crashes.

If the `card` is nil, the dictionary construction into `params.additionalAPIParameters`
crashes. This provides a slightly friendlier error message.

* Early exit test case if Card Source creation failed.

Previously, if the Card Source creation failed, the code crashed. Skipping the call to
[STPSourceParams threeDSecureParamsWithAmount:currency:returnURL:card:]` when card is nil
to prevent crashing.

Also calling `XCTFail()`, and fulfilling the 3DS expectation so that the test case doesn't
wait any longer than it needs to.

* Update expiration years from 2018 to 2024

* Add functional test for creating a card token with an expired year