Add tests for `STPRedirectContext initWithPaymentIntent:`
There are a bunch of safety checks in the init method to make sure this PaymentIntent
needs to be authorized via redirection, and that it parses as expected. If any of them
fail, the context should fail to be created.

Also verifies that creation succeeds and sets the right URLs & calls the completion
correctly.