Use `returnUrl` properties on `STPPaymentIntent` and `STPPaymentIntentParams`
Updates the `returnUrl` during confirmation to be correct for this app, and then
the `STPRedirectContext` can read the `returnUrl` off the Payment Intent object instead
of needing to pass it directly in.

Simplifies the communication with the backend, since it's now this client app talking
directly to Stripe to set the `returnUrl`, and from there it's handled for the developer.