Add support for STPPaymentIntent to STPRedirectContext (#987)
* Add support for STPPaymentIntent to STPRedirectContext

Added a new init method that takes the intent + return url.
Updated documentation

* Extract some STPRedirectContext properties into private header

Also moving STPPaymentIntent+Private header into the Project group with all the other
+Private headers.

* Using STPRedirectContext private url properties, add assertions to tests

These are the tests where the values of the URLs are important somehow, instead of checking
all the properties in all that test cases.

* Add explicit tests for `STPRedirectContext initWithSource:` using the newly exposed properties

This tests that construction pulls the right values out of the source. Each of these should
also be already tested below, implicitly, in the tests of the various workflows & states.

* Add fixture for STPPaymentIntent

* Add tests for `STPRedirectContext initWithPaymentIntent:`

There are a bunch of safety checks in the init method to make sure this PaymentIntent
needs to be authorized via redirection, and that it parses as expected. If any of them
fail, the context should fail to be created.

Also verifies that creation succeeds and sets the right URLs & calls the completion
correctly.

* Adding entry to MIGRATION.md about the renamed block

I don't think this will require a major version bump, since the symbol still exists and
is just deprecated.