Update `STPPaymentIntentFunctionalTest.testConfirmCanceledPaymentIntentFails`
Making it slightly less brittle to server-side changes. Previous test was failing
due to the addition of `requires_capture` in the list of acceptable statuses.

This is still asserting on a server-side string that might change, but I want some way
to make sure the error is the one I expect.