Use response from new `STPCustomer` object instead of previous one
Looks like a copy/paste error from f3b2d7250a50c1705c1776cb312f7a30676a86d2 in #864.

`STPCustomerContext` holds onto a `STPCustomer`. There are three places where it retrieves
a new `STPCustomer` - and when it does it needs to update the `sources` of that customer
based on the `STPCustomerContext.includeApplePaySources` boolean.

The bug is that we were using the *old* `STPCustomer.allResponseFields` (via
`self.customer`) instead of the `allResponseFields` property on the newly retrieved
`STPCustomer` object.

In isolation, it's a pretty obvious bug (once you know to look for it). There's no reason
we'd want to retrieve a Customer, and then use the response data for the previous object
to update the `sources`/`defaultSource` properties of the new one.

There is a fourth place that updates the sources, and that's inside
`setIncludeApplePaySources:`. In that case, we're updating the existing customer, and
using the existing customer's `allResponseFields`, so `self.customer.allResponseFields`
is correct there.