Change `STPCardParams.address` *back* to `strong` from `copy`
There are a bunch of places where we have `STPAddress` properties, and none of them
copy the object. Instead of changing `STPCardParams.address` as a one-off to be `copy`,
keep this `strong`.

I think the right thing to do would be to add a `STPMutableAddress` subclass, and take
a pass through the codebase. There are places where the STPAddress *should* be mutable,
like `STPCardParams.address`, allowing caller to build up the params to send to Stripe.

There are other places, like objects returned from the API, where it should be immutable.
Ex: `STPCard.address`.

This also makes the migrating instructions a lot simpler to understand.