Adding another copy call to `setCardParams:` and some documentation about the need for it
Per comment in `setCardParams:`, `setText:inField` calls trigger `paymentCardTextFieldDidChange:`
and `UIControlEventValueChanged`. This passes control back to the application code.

If the app code calls the `STPPaymentCardTextField.cardParams` property
accessor, the `internalCardParams` object is mutated to match the UI's state, and
that breaks the behavior of this method (motivating reason for this patch)

Similarly, if the app code changes the `STPCardParams` object that they own & passed into
this method during the delegate or value changed event, those changes could unexpectedly
show up.

The `desiredCardParams` copy is used so that at the end of the method's execution,
its done what was expected.