Change `STPPaymentCardTextField` to copy the `cardParams` object when it's set or read. (#1031)
* Adding tests showing issue from #1027, plus current behavior

* Conform STPAddress to NSCopying, with basic test to make sure it worked.

Using the runtime to iterate through all the properties and make sure they're equal.

* Conform STPCardParams to NSCopying, with basic test to make sure it worked.

Using the runtime to iterate through all the properties and make sure they're equal, or
not, as appropriate.

* Test showing that STPAddress objects are shared, instead of copied

* Test for name sharing between STPCardParams and its STPAddress

* Make `STPCardParams.address` property `copy` instead of `strong`

Also changed the `setName:`/`setAddress:` overrides slightly to avoid calling `setName:`
from `setAddress:`

* Make `STPPaymentCardTextField.cardParams` a copy property, instead of strong.

Fixes issue from #1027

* Change `STPPaymentCardTextField.cardParams` to return a copy, not `internalCardParams` ref

* update docs on `STPCardParams.address` re: name changing behavior

* Add some text to MIGRATING about breaking changes to `STPPaymentCardTextField` and `STPCardParams`

* Add objective-c language to older code blocks for better display

* Change `STPCardParams.address` *back* to `strong` from `copy`

There are a bunch of places where we have `STPAddress` properties, and none of them
copy the object. Instead of changing `STPCardParams.address` as a one-off to be `copy`,
keep this `strong`.

I think the right thing to do would be to add a `STPMutableAddress` subclass, and take
a pass through the codebase. There are places where the STPAddress *should* be mutable,
like `STPCardParams.address`, allowing caller to build up the params to send to Stripe.

There are other places, like objects returned from the API, where it should be immutable.
Ex: `STPCard.address`.

This also makes the migrating instructions a lot simpler to understand.

* Adding another copy call to `setCardParams:` and some documentation about the need for it

Per comment in `setCardParams:`, `setText:inField` calls trigger `paymentCardTextFieldDidChange:`
and `UIControlEventValueChanged`. This passes control back to the application code.

If the app code calls the `STPPaymentCardTextField.cardParams` property
accessor, the `internalCardParams` object is mutated to match the UI's state, and
that breaks the behavior of this method (motivating reason for this patch)

Similarly, if the app code changes the `STPCardParams` object that they own & passed into
this method during the delegate or value changed event, those changes could unexpectedly
show up.

The `desiredCardParams` copy is used so that at the end of the method's execution,
its done what was expected.

* Update hostname on SSL test domain with expired certificate.