Change Styleguide around acronym capitalization
The capitalization recommendation it made is *not* being followed by the app. I
specifically checked URL vs `Url`, and it's overwhelmingly `URL`. `stripeID` beats
`stripeId` by about 2 to 1. You can see fully capitalized acronyms in class names too.

I believe this better matches the actual style of the codebase, and will enable new code
to be consistent with the old.