Fix STPPaymentCardTextField nextFirstResponderField to never return `nil`
The type signature says `nonnull`, however it returns `nil` if index+1 is out of bounds.
This only happens if the postal code field is on, which most of our testing doesn't cover.
Instead, fall through to jump back to the first invalid field.

Also add docs to this method for what it returns.