Handle return key in STPPaymentCardTextField
This supports 'return' in any field to move to the next. If the STPPaymentCardTextField is
completely filled out and valid, and the user was in the last field, resigns first
responder instead.

Most users only see the return key in non-US countries in the postal code field. We use
the on-screen number pad keyboard for the CC number, expiration & CVC. However, with an
iPad or a hardware keyboard any field might receive the return key.