Fix a regression in v16 where the PaymentOptionsViewController no longer updates after a new card is added.
I'm not sure how common it is to use STPPaymentOptionsViewController in this way, but our UI Examples project does.
During the refactor for PaymentMethods, the code to refresh STPPaymentOptionsViewController once a new card is added
was dropped. This reimplements something similar using PaymentMethods. It also fixes the UI Examples project
to mock the appropriate PaymentMethod API call.