Don't query the service for non-UnionPay BINs (#63)
## Summary
Add a check to `retrieveBINRanges` to make sure users of STPBINRange can't unintentionally fetch metadata for non-UnionPay BINs.
While we're dealing with BINs, fix a compiler issue with `binRanges.map(\.brand)`. (Some users are reporting a compile error when Xcode can't infer the type from the key path: That doesn't make any sense to me, but this is the only place we use key-path expressions, so we might as well get rid of it.)

## Motivation
We were mistakenly querying the service in some situations with non-UnionPay BINs.

## Testing
Tested with Basic Integration using a 19-digit UnionPay test card.