Fix missing translations for language dialects (#92)
The old code didn't handle region-specific variations: If an app supported `fr-CH`, we'd compare `fr-CH` with `fr`, notice that they were different, and fall back to the app's main bundle. This would cause us to always show the unlocalized English key instead, as the string wouldn't be found.

But always reading the main bundle first means a user could accidentally override our strings in an unexpected way: For example, by replacing "Check" in a paper check payment with a word that means "✔️".

To work around this, we'll use the old behavior combined with glaures-ioki's change: If the languages don't match, we should check the main bundle for the string. If we don't find it there, we fall back to the Stripe bundle.

I also considered just checking the first two letters of the language code, but it's possible a user would want to override our implementation of a two-letter language prefix (for example, if we only supported `pt`, but a user required `pt-br`).