Fix STPTheme with navigation bar (#93)
We were updating the navigation bar appearance when the navigationController was nil. We'll update it after the view appears too, to make sure the theme takes effect. This was a pull request from rbenna.

While I was in this code, I fixed a bug where the Edit/Done button in the list of cards didn't always use the font from the theme.