MOBILESDK-264: Fix build settings to use xcconfig (#223)
Some settings in our `xcconfig` files were not being used because they were overwritten in the project file. This change updates the `.xcconfig` file to use the latest values and removes the overrides from the project file.

We'll be adding additional frameworks as part of our efforts to split up the SDK. To make maintaining multiple framework projects easier, we can reuse the same `.xcconfig` files across all SDK frameworks in the workspace – now any new framework project that is added can be configured to use the same `.xcconfig` files.