IDPROD-2546: Example app using native components (#384)
Added a `useNativeUI` property to `IdentityVerificationSheet` that uses native component screens instead of a webview. This property is only accessible via an SPI import and is not meant to be used externally. Its purpose is only to enable native components so we can easily test features during development.

Currently, enabling `useNativeUI` only displays an incomplete "Individual" screen of the IDV flow where the user inputs additional data (name, email, phone, etc). This is just so we can demo/test our form inputs, but eventually this will updated to return a NavigationViewController that navigates between all the screens of the IDV flow. 

Added an "Enable Native Components (experimental)" toggle to the playground in the example app which enables us to test with native components.