IDPROD-2455 part 3: ID Number TextFieldElement (#390)
* Adds ID Number field to section and updates number field type based on selected country
* New `TextFieldFormatter` formats strings based on `**####-###` shorthand notation where `*` represents letters and `#` numbers.
* New `IDNumberType` enum represents the various ID numbers that the client can accept and formats/validates them appropriately.
  * Note: These still need to be finalized, but eventually we'll pass string representations of this type from the backend to configure the accepted ID type per country
* Updates `TextFieldElement` to position the cursor relative to the end of the input instead of beginning. 
  * Previous to this change if the formatted text changed from `123` to `123-4`, the cursor would be placed after the `-` instead of the `4`.

Current behavior is that if an `IDNumberType` is specified, the input will show the number pad and require the length of the input match the format exactly. The only exception is Brazil CPF/CNPJ where the inputs can either be 11 or 14 characters (since CPF and CNPJ have different formats). If no `IDNumberType` is specified, then we will show the default keyboard with auto-capitolization and there are no restrictions on the inputs – we only require that there is at least one character if the field is non-optional.