IDPROD-2454: DateFieldElement for DOB input (#396)
Created a new `DateFieldElement` for date inputs and added a `dateOfBirth` factory method for Identity. This view works very similarly to our `DropdownFieldElement` except a date picker is displayed instead of a UIPicker and no chevron image is shown. Optionally, a defaultDate can be set on this element to pre-populate it with a date or it can be left blank until the user taps into the field and selects a date.

Specific changes include:
* Abstracted `DropdownFieldView` to take any UIView as its picker (e.g. UIDatePicker or UIPicker) and renamed it to `PickerFieldView`
* Created new `DateFieldElement` whose view is a `PickerFieldView`
* Added checks to `DropdownFieldElement` to ensure the defaultIndex is valid so we don't get an index out of bounds error
* Added DOB factory method to `IdentityElementsFactory`
* Added an `autosizeHeight` helper to `StripeCoreTestingUtils` to help layout views before taking snapshots