IDPROD-2545: Identity name section (#409)
Added a factory method for a name section that separates first and last name for identity verification.

This change abstracts `TextFieldElement.Address.NameConfiguration` so that it can be used for first name, last name, or full name fields. Also renamed `makeName()` to `makeFullName()` to disambiguate.

Also deleted internal initializers in for TextFieldConfigurations – because they're internal now, we can use the synthesized init rather than need to explicitly declare one.