IDPROD-2540 part 1: Make APIStubbedTestCase available in StripeCoreTestUtils (#442)
We're adding Identity endpoints and API models to the `StripeIdentity` module. We'd like to be able to write tests that utilize `APIStubbedTestCase` in `StripeIdentityTests`.

Prior to this change, `APIStubbedTestCase` existed in `StripeiOSTests`, which meant that no other module's tests had access to it. Now that it's been moved to `StripeCoreTestUtils`, all of our modules should be able to access this test class.

Also migrated `StripeCodableTests.swift` from `StripeiOSTests` -> `StripeCoreTests`. Since `StripeCodable` is in `StripeCore`, its test should exist in the parallel tests module. This also served as good case study to ensure that we could create tests dependent on `APIStubbedTestCase` outside of `StripeiOSTests`.