IDPROD-2541: VerificationPage endpoint tests + mock response (#447)
This includes tests for `postIdentityVerificationPage()`  and `VerificationSheetController` to ensure that requests are being properly made against the `VerificationPage` endpoint. This includes a new test helper, `MockData` which we can use to easily generate mock response objects from JSON.

Since this endpoint doesn't exist yet, I temporarily added the mock JSON file to the example app target and added the ability to specify a mock data fileURL in the example app so we could load a mock response for testing purposes. When the endpoint is live, we can remove this.