IDPROD-2739 part 1: Welcome and loading screen routing (#450)
Sets up infrastructure for local data persistence and screen routing between steps in the Verification Flow.

This doesn't contain all of the screens yet, but it currently displays a loading screen while waiting for the initial API response, then transitions to the Biometric Consent screen, followed by the AdditionalInfo screen:

https://user-images.githubusercontent.com/78050250/140444505-f336cba8-6a24-47a2-af3e-3d3522c2ea6a.mp4


This adds a few new classes:

#### VerificationSheetController
Responsible for making API requests and keeping track of the content returned by the API. Also responsible for telling the `VerificationSheetFlowController` when to load the first screen after the initial API request has returned.

#### VerificationSheetFlowController
Responsible for determining the next screen the user should see in the flow and updating the navigation stack. If any of the API calls returned an error, the flow-controller will (eventually) push an Error VC onto the stack.

#### VerificationSessionDataStore
Responsible for locally persisting user-input data. Any time the user finishes inputting data for a specific step in the flow, the `VerificationSheetController` saves the data from the `VerificationSessionDataStore` to our API endpoint.

#### IdentityFlowViewController
All screens in the flow will be represented by ViewControllers that subclass this. Each view controller is responsible for:
- Updating the `VerificationSessionDataStore` with data the user input on this screen
- Telling the `VerificationSheetController` to save the user's input data to our API
- Asking the `VerificationSheetFlowController` to transition to the next screen in the flow.

#### VerificationSheetAPIContent
Convenience data structure to join together the response from our two endpoints. This gets updated by the `VerificationSheetController` after each API call and passed to the `VerificationSheetFlowController` to determine the next step in the flow.


### Mocking

Additionally, some temporary code to help mock the `VerificationSessionData` API endpoint response has been put in place. This endpoint has a `required.missing` property that represents the required information the user still needs to input which then determines the next screen shown to the user. In order to make the screen routing demo-able, this naively checks the `VerificationSessionDataStore` for information already input by the user and removes it from `required.missing` before returning a response. The result is that we won't ever hit an error state while mocking the response, but it at least is enough to show the happy path for how this should work.